requirements.txt
asgiref==3.7.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.2
defusedxml==0.7.1
Deprecated==1.2.14
Django==4.2.3
django-braces==1.15.0
django-filter==22.1
django-oauth-toolkit==2.3.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
google-api-core==2.11.1
google-api-python-client==2.95.0
google-apis-oauth-django==1.0.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
httplib2==0.22.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.5.0
MarkupSafe==2.1.3
oauth2client==4.1.3
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.1
protobuf==4.23.4
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==1.7.1
pyparsing==3.1.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
wrapt==1.15.0
